change: replace the maplibre-compose UserLocationState to a simpler one
